Factor out Transformer into monad-transformer
Reduces dependency spaghetti in preparation for monad-gossip virtual
benchmarks.
