+++ monad-mock-swarm/benches/two_node_benchmark.rs
@@ -6,9 +6,9 @@ use monad_mock_swarm::{
     mock::{MockMempoolConfig, NoSerRouterConfig},
     mock_swarm::UntilTerminator,
     swarm_relation::NoSerSwarm,
-    transformer::{GenericTransformer, LatencyTransformer},
 };
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
+use monad_transformer::{GenericTransformer, LatencyTransformer};
 use monad_wal::mock::MockWALoggerConfig;
 
 pub fn criterion_benchmark(c: &mut Criterion) {
