+++ monad-mock-swarm/tests/common/mod.rs
@@ -7,14 +7,13 @@ use monad_crypto::NopSignature;
 use monad_executor::timed_event::TimedEvent;
 use monad_executor_glue::MonadEvent;
 use monad_gossip::mock::MockGossip;
-pub use monad_mock_swarm::swarm_relation::NoSerSwarm;
 use monad_mock_swarm::{
     mock::{MockMempool, MockMempoolConfig},
     swarm_relation::SwarmRelation,
-    transformer::BytesTransformerPipeline,
 };
 use monad_quic::{QuicRouterScheduler, QuicRouterSchedulerConfig};
 use monad_state::{MonadMessage, MonadState, VerifiedMonadMessage};
+use monad_transformer::BytesTransformerPipeline;
 use monad_validator::{simple_round_robin::SimpleRoundRobin, validator_set::ValidatorSet};
 use monad_wal::mock::{MockWALogger, MockWALoggerConfig};
