+++ monad-mock-swarm/tests/rand_lat.rs
@@ -1,14 +1,14 @@
 mod common;
 use std::env;
 
-use common::NoSerSwarm;
 use monad_consensus_types::transaction_validator::MockValidator;
 use monad_mock_swarm::{
     mock::{MockMempoolConfig, NoSerRouterConfig},
     mock_swarm::UntilTerminator,
-    transformer::GenericTransformer,
+    swarm_relation::NoSerSwarm,
 };
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
+use monad_transformer::GenericTransformer;
 use monad_wal::mock::MockWALoggerConfig;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use test_case::test_case;
@@ -51,7 +51,7 @@ fn nodes_with_random_latency_cron() {
 fn nodes_with_random_latency(seed: u64) {
     use std::time::Duration;
 
-    use monad_mock_swarm::transformer::RandLatencyTransformer;
+    use monad_transformer::RandLatencyTransformer;
 
     create_and_run_nodes::<NoSerSwarm, _, _>(
         MockValidator,
