+++ monad-mock-swarm/tests/random_mempool.rs
@@ -14,10 +14,10 @@ use monad_mock_swarm::{
     },
     mock_swarm::UntilTerminator,
     swarm_relation::SwarmRelation,
-    transformer::{GenericTransformer, GenericTransformerPipeline, LatencyTransformer},
 };
 use monad_state::{MonadMessage, MonadState, VerifiedMonadMessage};
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
+use monad_transformer::{GenericTransformer, GenericTransformerPipeline, LatencyTransformer};
 use monad_validator::{simple_round_robin::SimpleRoundRobin, validator_set::ValidatorSet};
 use monad_wal::mock::{MockWALogger, MockWALoggerConfig};
