+++ monad-mock-swarm/tests/single_node.rs
@@ -2,16 +2,17 @@ mod common;
 
 use std::time::{Duration, Instant};
 
-use common::{NoSerSwarm, QuicSwarm};
+use common::QuicSwarm;
 use monad_consensus_types::transaction_validator::MockValidator;
 use monad_gossip::mock::MockGossipConfig;
 use monad_mock_swarm::{
     mock::{MockMempoolConfig, NoSerRouterConfig},
     mock_swarm::UntilTerminator,
-    transformer::{BwTransformer, BytesTransformer, GenericTransformer, LatencyTransformer},
+    swarm_relation::NoSerSwarm,
 };
 use monad_quic::QuicRouterSchedulerConfig;
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
+use monad_transformer::{BwTransformer, BytesTransformer, GenericTransformer, LatencyTransformer};
 use monad_wal::mock::MockWALoggerConfig;
 use tracing_test::traced_test;
