+++ monad-testutil/src/swarm.rs
@@ -18,9 +18,9 @@ use monad_mock_swarm::{
     mock::MockExecutor,
     mock_swarm::{Node, Nodes, NodesTerminator},
     swarm_relation::SwarmRelation,
-    transformer::ID,
 };
 use monad_state::MonadConfig;
+use monad_transformer::ID;
 use monad_types::NodeId;
 
 use crate::{signing::get_genesis_config, validators::create_keys_w_validators};
