+++ monad-transformer/Cargo.toml
@@ -1,20 +1,19 @@
 [package]
-name = "monad-gossip"
+name = "monad-transformer"
 version = "0.1.0"
 edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 # this is necessary for criterion bench options
 # https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options
 [lib]
 bench = false
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-monad-crypto = { path = "../monad-crypto" }
-monad-transformer = { path = "../monad-transformer" }
+monad-tracing-counter = { path = "../monad-tracing-counter" }
 monad-types = { path = "../monad-types" }
-bincode = "1.3"
+
 rand = { workspace = true }
 rand_chacha = { workspace = true }
-serde = "1.0"
+tracing = { workspace = true }
