+++ monad-virtual-bench/benches/many_nodes_bench.rs
@@ -14,11 +14,13 @@ use monad_mock_swarm::{
     mock::{MockMempool, MockMempoolConfig},
     mock_swarm::UntilTerminator,
     swarm_relation::SwarmRelation,
-    transformer::{BwTransformer, BytesTransformer, BytesTransformerPipeline, LatencyTransformer},
 };
 use monad_quic::{QuicRouterScheduler, QuicRouterSchedulerConfig};
 use monad_state::{MonadMessage, MonadState, VerifiedMonadMessage};
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
+use monad_transformer::{
+    BwTransformer, BytesTransformer, BytesTransformerPipeline, LatencyTransformer,
+};
 use monad_types::{NodeId, Round};
 use monad_validator::{simple_round_robin::SimpleRoundRobin, validator_set::ValidatorSet};
 use monad_wal::mock::{MockWALogger, MockWALoggerConfig};
