+++ monad-viz/Cargo.toml
@@ -18,6 +18,7 @@ monad-executor-glue = { path = "../monad-executor-glue" }
 monad-mock-swarm = { path = "../monad-mock-swarm" }
 monad-state = { path = "../monad-state" }
 monad-testutil = { path = "../monad-testutil" }
+monad-transformer = { path = "../monad-transformer" }
 monad-types = { path = "../monad-types" }
 monad-validator = { path = "../monad-validator" }
 monad-wal = { path = "../monad-wal" }
