+++ monad-viz/src/main.rs
@@ -35,12 +35,11 @@ use monad_executor_glue::MonadEvent;
 use monad_mock_swarm::{
     mock::{MockMempool, MockMempoolConfig, NoSerRouterConfig, NoSerRouterScheduler},
     swarm_relation::SwarmRelation,
-    transformer::{
-        GenericTransformer, GenericTransformerPipeline, LatencyTransformer, XorLatencyTransformer,
-        ID,
-    },
 };
 use monad_state::{MonadMessage, MonadState, VerifiedMonadMessage};
+use monad_transformer::{
+    GenericTransformer, GenericTransformerPipeline, LatencyTransformer, XorLatencyTransformer, ID,
+};
 use monad_types::NodeId;
 use monad_validator::{simple_round_robin::SimpleRoundRobin, validator_set::ValidatorSet};
 use monad_wal::{
